Remove Content-Security-Policy on specific assets
