+++ config/initializers/secure_headers.rb
@@ -69,22 +69,35 @@ SecureHeaders::Configuration.default do |config| # rubocop:disable Metrics/Block
   # }
 end
 
+# A tiny middleware that calls a block on each request.
+# If the block returns true, it deletes the Content-Security-Policy header
+# Intended so that we can override SecureHeaders behavior and not set
+# the headers on asset files
+class SecureHeaders::RemoveContentSecurityPolicy
+  # @yieldparam [Rack::Request] request
+  def initialize(app, &block)
+    @app = app
+    @block = block
+  end
+
+  def call(env)
+    status, headers, body = @app.call(env)
+
+    should_remove = @block.call(Rack::Request.new(env))
+    headers.delete('Content-Security-Policy') if should_remove
+
+    [status, headers, body]
+  end
+end
 
 # We need this to be called after the SecureHeaders::Railtie adds its own middleware at the top
 Rails.application.configure do |config|
-  require 'manual_secure_headers_override'
-
   acuant_sdk_static_files = %w[
     AcuantImageProcessingWorker.min.js
     AcuantImageProcessingWorker.wasm
   ].freeze
 
-  config.middleware.insert_after 0, ManualSecureHeadersOverride do |request|
-    if acuant_sdk_static_files.any? { |file| request.path.end_with?(file) }
-      SecureHeaders.append_content_security_policy_directives(
-        request,
-        script_src: ["'unsafe-eval'"]
-      )
-    end
+  config.middleware.insert_before 0, SecureHeaders::RemoveContentSecurityPolicy do |request|
+    acuant_sdk_static_files.any? { |file| request.path.end_with?(file) }
   end
 end
